Split goDecorator service into two services
